Working with new system
New system eliminates the buggy, complicated ros sphero driver I
downloaded and replaces it with a simple but non-encapsulated driver
that just sends to the backend driver whatever it receives over the ROS
stream. This is useful since it's easy to do anything the backend driver
can do but the ROS driver maintains the bluetooth connection.